Added views to coach profile
