--- app/views/coach_profile/edit.html.erb
@@ -1,3 +1,0 @@
-<%= render "form", coach_profile: @coach_profile %>
-
-<%= link_to 'lessons', lessons_path, class: "btn btn-primary" %>
