--- app/views/coach_profile/show.html.erb
@@ -1,10 +1,0 @@
-<h1>Coach Profile Page</h1>
-
-<h2><%= @user.name %></h2>
-<h2><%= @coach_profile.description %></h2>
-<h2><%= @coach_profile.sport_id %></h2>
-
-
-<%= link_to 'edit', edit_user_coach_profile, class: "btn btn-primary"  %>
-
-<%= link_to 'lessons', lessons_path, class: "btn btn-primary" %>
