+++ app/views/coach_profiles/edit.html.erb
@@ -1,0 +1,2 @@
+<%= render "form", coach_profile: @coach_profile, user: @user %>
+
