+++ app/views/coach_profiles/new.html.erb
@@ -1,0 +1,1 @@
+<%= render "form", coach_profile: @coach_profile, user: @user %>
