+++ app/views/coach_profiles/show.html.erb
@@ -1,0 +1,10 @@
+<h1>Coach Profile Page</h1>
+
+<h2>User name: <%= @user.name %></h2>
+<h2>Coach profile description: <%= @coach_profile.description %></h2>
+<h2>Coach's sport: <%= @coach_profile.sport.name %></h2>
+
+
+<%= link_to 'edit', edit_user_coach_profile_path, class: "btn btn-primary"  %>
+
+<%= link_to 'lessons', lessons_path, class: "btn btn-primary" %>
